port(phase2): joints, beer and хлам are signed words, as the record says
`20ae:38c5`, `20ae:38c3` and `20ae:38c9` are Pascal `Integer`s -- signed
16-bit words -- and every gate on them is a SIGNED compare:

    1000:e9aa  cmp word [0x38c5],0x0 / jnle    the joint
    1000:db33  cmp word [0x38c3],0x0 / jle     the beer
    1000:ce87  cmp word [0x38c9],0x0 / jle     the хлам

`src/save.rs` already declared all three `i16`, which is the record's own
width; `Fighter` held them as `u16`. That mismatch cost a conversion in each
direction and lost information in both:

* `to_save` cast `u16 -> i16` silently;
* `from_save` did `.max(0) as u16`, CLAMPING a negative word to zero. The
  original keeps it. `tests/save_load.rs` had that clamp pinned as expected
  behaviour, with a `-300` хлам in the fixture.

`Fighter::joints`, `beer_dl` and `junk` are now `i16`. Both conversions are
gone -- `to_save` assigns and `from_save` assigns -- and for these three the
`Game -> Save -> Game` round trip is the identity, which the updated test
asserts in both directions.

Two gates stop approximating. `Game::smoke` read `joints == 0` where
`1000:e9af jnle` means "strictly positive", which agreed for every
non-negative value and could not express the rest; it is `<= 0` now.
`Game::sell_junk` and the beer gate carried `(x as i16) <= 0` casts to get
the signed read -- the field is signed, so the casts are gone.

`dealers_x_refuses_a_junk_word_whose_top_bit_is_set` now sets `i16::MIN`
directly instead of `0x8000`, which is what that word always meant.

The oracles needed no regeneration and are not touched: `combat_trace.json`
records `e_money_396c` and friends with `"width": 2` and `e_armor_3968` with
`"width": 1`, so the capture has always carried the ORIGINAL's widths. The
fixtures that read them now reinterpret (`as i16`) rather than convert, and
`tests/combat_vectors.rs` is unaffected -- it holds no purse fields at all.

`cargo test` 460 passed, 0 failed. `difftest` 363/363. clippy and fmt clean.

+++ tests/combat_sequence.rs
@@ -285,9 +285,11 @@ fn game_for(run: &Run) -> Game {
         broken_leg: b[0x215] != 0,
         armor: u16::from(b[0x216]),
         money: i32::from(u16at(0x22b)), // 20ae:38c7
-        beer_dl: u16at(0x227),          // 20ae:38c3
-        junk: u16at(0x22d),             // 20ae:38c9
-        joints: u16at(0x229),           // 20ae:38c5
+        beer_dl: u16at(0x227) as i16,   // 20ae:38c3
+        junk: u16at(0x22d) as i16,      // 20ae:38c9
+        // 20ae:38c5 is a signed Integer; the bit pattern is what the record
+        // holds, so the read is a reinterpretation, not a conversion.
+        joints: u16at(0x229) as i16, // 20ae:38c5
         ..Fighter::default()
     };
     // `array[1..40] of string[2]` at `.SAV 0x236` (`20ae:38d2`, reached
@@ -489,9 +491,15 @@ fn replay(label: &str) -> Game {
         assert_eq!(got.broken_jaw, w.e_broken_jaw_3966 != 0, "{at}: 20ae:3966");
         assert_eq!(got.broken_leg, w.e_broken_leg_3967 != 0, "{at}: 20ae:3967");
         assert_eq!(got.armor, w.e_armor_3968, "{at}: 20ae:3968");
-        assert_eq!(got.beer_dl, w.e_beer_396a, "{at}: 20ae:396a loot beer");
+        // `20ae:396a` and `20ae:396e` are signed Integers in the record
+        // (`combat_trace.json` calls both width 2); the capture stores the
+        // bit pattern, so this reinterprets rather than converts.
+        assert_eq!(
+            got.beer_dl, w.e_beer_396a as i16,
+            "{at}: 20ae:396a loot beer"
+        );
         assert_eq!(got.money, i32::from(w.e_money_396c), "{at}: 20ae:396c");
-        assert_eq!(got.junk, w.e_hlam_396e, "{at}: 20ae:396e Хлам");
+        assert_eq!(got.junk, w.e_hlam_396e as i16, "{at}: 20ae:396e Хлам");
     }
 
     // The per-round channel: one entry per 1000:441d stop, both fighters'
@@ -615,13 +623,16 @@ fn assert_final_state(label: &str, run: &Run, g: &Game) {
     // The purse. `1000:523e`..`1000:5251` is what fills it on a win, and this
     // port did not reproduce that block until Task 13 -- `docs/re/gaps.md`
     // recorded it as the reason `Fighter::junk` stayed 0.
-    assert_eq!(g.player.beer_dl, f.beer_38c3, "{label}: 20ae:38c3 beer");
+    assert_eq!(
+        g.player.beer_dl, f.beer_38c3 as i16,
+        "{label}: 20ae:38c3 beer"
+    );
     assert_eq!(
         u16::try_from(g.player.money).expect("money fits a word"),
         f.money_38c7,
         "{label}: 20ae:38c7 money"
     );
-    assert_eq!(g.player.junk, f.hlam_38c9, "{label}: 20ae:38c9 Хлам");
+    assert_eq!(g.player.junk, f.hlam_38c9 as i16, "{label}: 20ae:38c9 Хлам");
     assert_eq!(
         g.rng.state(),
         f.randseed_367e,
